+++ src/MicroElements.Metadata/Metadata/EmptyPropertyContainer.cs
@@ -2,6 +2,7 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using System;
+using System.Collections;
 using System.Collections.Generic;
 
 namespace MicroElements.Metadata
@@ -27,9 +28,30 @@ namespace MicroElements.Metadata
         public IReadOnlyList<IPropertyValue> Properties => Array.Empty<IPropertyValue>();
 
         /// <inheritdoc />
-        public T GetValue<T>(IProperty<T> property) => property.DefaultValue();
+        public IPropertyValue<T> GetPropertyValue<T>(IProperty<T> property, PropertySearch propertySearch = PropertySearch.Default) => new PropertyValue<T>(property, default, ValueSource.NotDefined);
+
+        /// <inheritdoc />
+        public IPropertyValue GetPropertyValueUntyped(IProperty property, PropertySearch propertySearch = PropertySearch.Default) => null;
+
+        /// <inheritdoc />
+        public T GetValue<T>(IProperty<T> property, PropertySearch propertySearch = PropertySearch.Default) => property.DefaultValue();
 
         /// <inheritdoc />
         public object GetValueUntyped(IProperty property) => null;
+
+        /// <inheritdoc />
+        public IEnumerator<IPropertyValue> GetEnumerator()
+        {
+            yield break;
+        }
+
+        /// <inheritdoc />
+        IEnumerator IEnumerable.GetEnumerator() => GetEnumerator();
+
+        /// <inheritdoc />
+        public int Count => 0;
+
+        /// <inheritdoc />
+        public IPropertyValue this[int index] => null;
     }
 }