+++ src/MicroElements.Metadata/Metadata/IPropertyContainer.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) MicroElements. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
+using System;
 using System.Collections.Generic;
 
 namespace MicroElements.Metadata
@@ -8,7 +9,7 @@ namespace MicroElements.Metadata
     /// <summary>
     /// Represents object that contains properties and values for these properties.
     /// </summary>
-    public interface IPropertyContainer
+    public interface IPropertyContainer : IReadOnlyList<IPropertyValue>
     {
         /// <summary>
         /// Gets parent property source.
@@ -21,12 +22,30 @@ namespace MicroElements.Metadata
         IReadOnlyList<IPropertyValue> Properties { get; }
 
         /// <summary>
+        /// Gets property value.
+        /// </summary>
+        /// <typeparam name="T">Property type.</typeparam>
+        /// <param name="property">Property to find.</param>
+        /// <param name="propertySearch">Search conditions.</param>
+        /// <returns>Property value or null.</returns>
+        IPropertyValue<T> GetPropertyValue<T>(IProperty<T> property, PropertySearch propertySearch = PropertySearch.Default);
+
+        /// <summary>
+        /// Gets property value.
+        /// </summary>
+        /// <param name="property">Property to find.</param>
+        /// <param name="propertySearch">Search conditions.</param>
+        /// <returns>Property value or null.</returns>
+        IPropertyValue GetPropertyValueUntyped(IProperty property, PropertySearch propertySearch = PropertySearch.Default);
+
+        /// <summary>
         /// Gets the value for property.
         /// </summary>
         /// <typeparam name="T">Property type.</typeparam>
         /// <param name="property">Property to find.</param>
+        /// <param name="propertySearch">Search conditions.</param>
         /// <returns>The value for property.</returns>
-        T GetValue<T>(IProperty<T> property);
+        T GetValue<T>(IProperty<T> property, PropertySearch propertySearch = PropertySearch.Default);
 
         /// <summary>
         /// Gets untyped value for property.
@@ -39,7 +58,7 @@ namespace MicroElements.Metadata
     /// <summary>
     /// Property container that can mutate it state.
     /// </summary>
-    public interface IMutablePropertyContainer
+    public interface IMutablePropertyContainer : IPropertyContainer
     {
         /// <summary>
         /// Sets value for property.
@@ -49,5 +68,21 @@ namespace MicroElements.Metadata
         /// <param name="value">Value to store.</param>
         /// <returns><see cref="IPropertyValue{T}"/> that holds value for property.</returns>
         IPropertyValue<T> SetValue<T>(IProperty<T> property, T value);
+    }
+
+    /// <summary>
+    /// Search conditions.
+    /// </summary>
+    [Flags]
+    public enum PropertySearch
+    {
+        Default = Equals | SearchInParent,
+        All = Equals | ByName | ByAlias | IgnoreCase | SearchInParent,
+
+        Equals = 1,
+        ByName = 2,
+        ByAlias = 4,
+        IgnoreCase = 8,
+        SearchInParent = 16
     }
 }