+++ src/MicroElements.Metadata/Metadata/PropertyContainerExtensions.cs
@@ -18,42 +18,37 @@ namespace MicroElements.Metadata
         /// </summary>
         /// <param name="propertyContainer">Property container.</param>
         /// <param name="property">Property to search.</param>
-        /// <param name="searchInParent">Search property in parent source if not found in container.</param>
+        /// <param name="search">Search conditions.</param>
         /// <returns>True if property exists in container.</returns>
-        public static bool HasValue(this IPropertyContainer propertyContainer, IProperty property, bool searchInParent = true)
+        public static bool HasValue(this IPropertyContainer propertyContainer, IProperty property, PropertySearch search = PropertySearch.Default)
         {
-            if (propertyContainer.Properties.ContainsProperty(property))
-                return true;
-
-            if (searchInParent && propertyContainer.ParentSource.Properties.Count > 0 && propertyContainer.ParentSource.HasValue(property, true))
-                return true;
-
-            return false;
+            return propertyContainer.GetPropertyValueUntyped(property, search) != null;
         }
 
-        /// <summary>
-        /// Returns true if <paramref name="propertyList"/> contains property by name or alias.
-        /// </summary>
-        /// <param name="propertyList">Property list.</param>
-        /// <param name="propertyName">Property name to search.</param>
-        /// <returns>True if property found in list.</returns>
-        public static bool ContainsPropertyByNameOrAlias(this IEnumerable<IPropertyValue> propertyList, string propertyName)
-            => propertyList.Any(propertyValue => IsMatchesByNameOrAlias(propertyValue, propertyName));
+        public static void SetValueIfNotSet<T>(this IMutablePropertyContainer propertyContainer, IProperty<T> property, T value, PropertySearch search = PropertySearch.Equals)
+        {
+            var propertyValue = propertyContainer.GetPropertyValue(property, search);
+            if (propertyValue == null)
+                propertyContainer.SetValue(property, value);
+        }
 
         public static IPropertyValue GetPropertyByNameOrAlias(this IEnumerable<IPropertyValue> propertyList, string propertyName)
             => propertyList.FirstOrDefault(propertyValue => IsMatchesByNameOrAlias(propertyValue, propertyName));
 
-        public static bool ContainsPropertyByNameOrAlias(this IEnumerable<IPropertyValue> propertyList, IProperty property)
-            => propertyList.Any(propertyValue => IsMatchesByNameOrAlias(propertyValue, property));
+        public static IPropertyValue GetProperty(this IReadOnlyList<IPropertyValue> propertyList, IProperty property, PropertySearch propertySearch = PropertySearch.All)
+        {
+            IPropertyValue result = null;
+            if (propertySearch.HasFlag(PropertySearch.Equals))
+                result = propertyList.FirstOrDefault(propertyValue => propertyValue.PropertyUntyped == property);
+            if (result == null && propertySearch.HasFlag(PropertySearch.ByName))
+                result = propertyList.FirstOrDefault(propertyValue => propertyValue.PropertyUntyped.Name.IsOneOf(property.Name, property.Alias, propertySearch));
+            if (result == null && propertySearch.HasFlag(PropertySearch.ByAlias))
+                result = propertyList.FirstOrDefault(propertyValue => propertyValue.PropertyUntyped.Alias.IsOneOf(property.Name, property.Alias, propertySearch));
+            return result;
+        }
 
         public static bool ContainsProperty(this IEnumerable<IPropertyValue> propertyList, IProperty property)
             => propertyList.Any(propertyValue => propertyValue.PropertyUntyped == property);
-
-        public static bool IsMatchesByNameOrAlias(this IPropertyValue propertyValue, IProperty property)
-        {
-            return propertyValue.IsMatchesByNameOrAlias(property.Name) ||
-                   propertyValue.IsMatchesByNameOrAlias(property.Alias);
-        }
 
         public static bool IsMatchesByNameOrAlias(this IPropertyValue propertyValue, string propertyName)
         {
@@ -64,5 +59,18 @@ namespace MicroElements.Metadata
 
         private static bool EqualsIgnoreCase(this string value, string other) =>
             value?.Equals(other, StringComparison.OrdinalIgnoreCase) ?? other == null;
+
+        private static bool IsOneOf(this string value, string other1, string other2, PropertySearch search)
+        {
+            static StringComparison ComparisonByFlag(PropertySearch propertySearch) =>
+                propertySearch.HasFlag(PropertySearch.IgnoreCase)
+                    ? StringComparison.OrdinalIgnoreCase
+                    : StringComparison.Ordinal;
+
+            if (value == null)
+                return other1 == null || other2 == null;
+
+            return value.Equals(other1, ComparisonByFlag(search)) || value.Equals(other2, ComparisonByFlag(search));
+        }
     }
 }